Split the code into header and source files, edited audioglyph.make
